+++ plugins/modules/dns_zone.py
@@ -7,10 +7,11 @@ from ansible import __version__
 from ansible.module_utils.basic import AnsibleModule, env_fallback
 from ansible.module_utils._text import to_native
 import re
+import uuid
 
 HAS_SDK = True
 try:
-    import ionoscloud_container_registry
+    import ionoscloud_dns
 except ImportError:
     HAS_SDK = False
 
@@ -20,37 +21,34 @@ ANSIBLE_METADATA = {
     'supported_by': 'community',
 }
 
-CONTAINER_REGISTRY_USER_AGENT = 'ansible-module/%s_ionos-cloud-sdk-python-container-registry/%s'% (
-    __version__, ionoscloud_container_registry.__version__,
+DNS_USER_AGENT = 'ansible-module/%s_ionos-cloud-sdk-python-dns/%s'% (
+    __version__, ionoscloud_dns.__version__,
 )
-DOC_DIRECTORY = 'container-registry'
+DOC_DIRECTORY = 'dns'
 STATES = ['present', 'absent', 'update']
-OBJECT_NAME = 'Registry'
-RETURNED_KEY = 'registry'
+OBJECT_NAME = 'Zone'
+RETURNED_KEY = 'zone'
+REPO_URL = "https://github.com/ionos-cloud/module-ansible"
 
 OPTIONS = {
-    'garbage_collection_schedule': {
-        'description': [
-            'Dict containing "time" (the time of the day when to perform the garbage_collection) '
-            'and "days" (the days when to perform the garbage_collection).',
-        ],
+    'enabled': {
+        'description': ['Users can activate and deactivate zones.'],
         'available': ['present', 'update'],
-        'type': 'dict',
+        'type': 'bool',
     },
-    'location': {
-        'description': ['The location of your registry'],
+    'description': {
+        'description': ['The hosted zone is used for...'],
         'available': ['present', 'update'],
-        'required': ['present'],
         'type': 'str',
     },
     'name': {
-        'description': ['The name of your registry.'],
+        'description': ['The zone name'],
         'available': ['present', 'update'],
         'required': ['present'],
         'type': 'str',
     },
-    'registry': {
-        'description': ['The ID or name of an existing Registry.'],
+    'zone': {
+        'description': ['The ID or name of an existing Zone.'],
         'available': ['update', 'absent'],
         'required': ['update', 'absent'],
         'type': 'str',
@@ -119,11 +117,6 @@ OPTIONS = {
     },
 }
 
-IMMUTABLE_OPTIONS = [
-    { "name": "name", "note": "" },
-    { "name": "location", "note": "" },
-]
-
 def transform_for_documentation(val):
     val['required'] = len(val.get('required', [])) == len(STATES)
     del val['available']
@@ -133,10 +126,10 @@ def transform_for_documentation(val):
 
 DOCUMENTATION = '''
 ---
-module: registry
-short_description: Allows operations with Ionos Cloud Registries.
+module: dns_zone
+short_description: Allows operations with Ionos Cloud DNS Zones.
 description:
-     - This is a module that supports creating, updating or destroying Registries
+     - This is a module that supports creating, updating or destroying DNS Zones
 version_added: "2.0"
 options:
 ''' + '  ' + yaml.dump(
@@ -145,35 +138,30 @@ options:
 requirements:
     - "python >= 2.6"
     - "ionoscloud >= 6.0.2"
-    - "ionoscloud-container-registry >= 1.0.0"
+    - "ionoscloud-dns >= 1.0.0"
 author:
     - "IONOS Cloud SDK Team <sdk-tooling@ionos.com>"
 '''
 
 EXAMPLE_PER_STATE = {
-    'present': '''- name: Create Registry
-    registry:
-      name: test_registry
-      location: de/fra
-      garbage_collection_schedule:
-        days: 
-            - Wednesday
-        time: 04:17:00+00:00
-    register: registry_response
+    'present': '''- name: Create Zone
+    dns_zone:
+      name: example.com
+      description: zone_description
+      enabled: true
+    register: zone_response
   ''',
-    'update': '''- name: Update Registry
-    registry:
-      registry: test_registry
-      name: test_registry_update
-      garbage_collection_schedule:
-        days: 
-            - Wednesday
-        time: 04:17:00+00:00
-    register: updated_registry_response
+    'update': '''- name: Update zone
+    dns_zone:
+      zone: example.com
+      description: zone_description_update
+      enabled: false
+      state: update
+    register: updated_zone_response
   ''',
-    'absent': '''- name: Delete Registry
-    registry:
-      registry: test_registry
+    'absent': '''- name: Delete zone
+    dns_zone:
+      zone: example.com
       wait: true
       state: absent
   ''',
@@ -222,29 +210,22 @@ def get_resource_id(module, resource_list, identity, identity_paths=None):
 
 
 def _should_replace_object(module, existing_object):
-    return (
-        module.params.get('location') is not None
-        and existing_object.properties.location != module.params.get('location')
-        or module.params.get('name') is not None
-        and existing_object.properties.name != module.params.get('name')
-    )
+    return False
 
 
 def _should_update_object(module, existing_object):
-    gc_schedule = module.params.get('garbage_collection_schedule')
     return (
-        gc_schedule is not None
-        and (
-            gc_schedule.get('days') is not None
-            and sorted(existing_object.properties.garbage_collection_schedule.days) != sorted(gc_schedule.get('days'))
-            or gc_schedule.get('time') is not None
-            and existing_object.properties.garbage_collection_schedule.time != gc_schedule.get('time')
-        )
+        module.params.get('description') is not None
+        and existing_object.properties.description != module.params.get('description')
+        or module.params.get('enabled') is not None
+        and existing_object.properties.enabled != module.params.get('enabled')
+        or module.params.get('name') is not None
+        and existing_object.properties.zone_name != module.params.get('name')
     )
 
 
 def _get_object_list(module, client):
-    return ionoscloud_container_registry.RegistriesApi(client).registries_get()
+    return ionoscloud_dns.ZonesApi(client).zones_get()
 
 
 def _get_object_name(module):
@@ -252,84 +233,64 @@ def _get_object_name(module):
 
 
 def _get_object_identifier(module):
-    return module.params.get('registry')
+    return module.params.get('zone')
 
 
 def _create_object(module, client, existing_object=None):
-    gc_schedule = module.params.get('garbage_collection_schedule')
-    if gc_schedule:
-        gc_schedule = ionoscloud_container_registry.WeeklySchedule(
-            days=gc_schedule.get('days'),
-            time=gc_schedule.get('time'),
-        )
     name = module.params.get('name')
-    location = module.params.get('location')
+    description = module.params.get('description')
+    enabled = module.params.get('enabled')
     if existing_object is not None:
         name = existing_object.properties.name if name is None else name
-        location = existing_object.properties.location if location is None else location
-        gc_schedule = existing_object.properties.garbage_collection_schedule if gc_schedule is None else gc_schedule
+        description = existing_object.properties.description if description is None else description
+        enabled = existing_object.properties.enabled if enabled is None else enabled
 
-    registries_api = ionoscloud_container_registry.RegistriesApi(client)
-
-    registry_properties = ionoscloud_container_registry.PostRegistryProperties(
-        name=name,
-        location=location,
-        garbage_collection_schedule=gc_schedule,
+    zones_api = ionoscloud_dns.ZonesApi(client)
+    zone = ionoscloud_dns.ZoneEnsure(
+        properties=ionoscloud_dns.Zone(
+            zone_name=name,
+            description=description,
+            enabled=enabled,
+        ),
     )
 
-    registry = ionoscloud_container_registry.PostRegistryInput(properties=registry_properties)
-
     try:
-        registry = registries_api.registries_post(registry)
-    except ionoscloud_container_registry.ApiException as e:
-        module.fail_json(msg="failed to create the new Registry: %s" % to_native(e))
-    return registry
+        zone = zones_api.zones_put(str(uuid.uuid5(uuid.uuid5(uuid.NAMESPACE_URL, REPO_URL), str(uuid.uuid4()))), zone)
+    except ionoscloud_dns.ApiException as e:
+        module.fail_json(msg="failed to create the new DNS Zone: %s" % to_native(e))
+    return zone
 
 
 def _update_object(module, client, existing_object):
-    gc_schedule = module.params.get('garbage_collection_schedule')
-    if gc_schedule:
-        gc_schedule = ionoscloud_container_registry.WeeklySchedule(
-            days=gc_schedule.get('days'),
-            time=gc_schedule.get('time'),
-        )
+    name = module.params.get('name')
+    description = module.params.get('description')
+    enabled = module.params.get('enabled')
+    if existing_object is not None:
+        name = existing_object.properties.zone_name if name is None else name
+        description = existing_object.properties.description if description is None else description
+        enabled = existing_object.properties.enabled if enabled is None else enabled
 
-    registries_api = ionoscloud_container_registry.RegistriesApi(client)
-
-    registry_properties = ionoscloud_container_registry.PatchRegistryInput(
-        garbage_collection_schedule=gc_schedule,
-    )
+    zone = ionoscloud_dns.ZoneEnsure(properties=ionoscloud_dns.Zone(
+        zone_name=name,
+        description=description,
+        enabled=enabled,
+    ))
 
     try:
-        registry = registries_api.registries_patch(
-            registry_id=existing_object.id,
-            patch_registry_input=registry_properties,
-        )
+        zone = ionoscloud_dns.ZonesApi(client).zones_put(zone_id=existing_object.id, zone_ensure=zone)
 
-        return registry
-    except ionoscloud_container_registry.ApiException as e:
-        module.fail_json(msg="failed to update the Registry: %s" % to_native(e))
+        return zone
+    except ionoscloud_dns.ApiException as e:
+        module.fail_json(msg="failed to update the DNS Zone: %s" % to_native(e))
 
 
 def _remove_object(module, client, existing_object):
-    registries_api = ionoscloud_container_registry.RegistriesApi(client)
-    names_api = ionoscloud_container_registry.NamesApi(client)
+    zones_api = ionoscloud_dns.ZonesApi(client)
 
     try:
-        registries_api.registries_delete(existing_object.id)
-
-        if module.params.get('wait'):
-            try:
-                client.wait_for(
-                    fn_request=lambda: names_api.names_check_usage(existing_object.properties.name),
-                    fn_check=lambda _: False,
-                    scaleup=10000,
-                )
-            except ionoscloud_container_registry.ApiException as e:
-                if e.status != 404:
-                    raise e
-    except ionoscloud_container_registry.ApiException as e:
-        module.fail_json(msg="failed to remove the Registry: %s" % to_native(e))
+        zones_api.zones_delete(existing_object.id)
+    except ionoscloud_dns.ApiException as e:
+        module.fail_json(msg="failed to remove the DNS Zone: %s" % to_native(e))
 
 
 def update_replace_object(module, client, existing_object):
@@ -365,7 +326,10 @@ def update_replace_object(module, client, existing_object):
 
 
 def create_object(module, client):
-    existing_object = get_resource(module, _get_object_list(module, client), _get_object_name(module))
+    existing_object = get_resource(
+        module, _get_object_list(module, client), _get_object_name(module),
+        identity_paths=[['id'], ['properties', 'zone_name']],
+    )
 
     if existing_object:
         return update_replace_object(module, client, existing_object)
@@ -382,13 +346,18 @@ def update_object(module, client):
     object_name = _get_object_name(module)
     object_list = _get_object_list(module, client)
 
-    existing_object = get_resource(module, object_list, _get_object_identifier(module))
+    existing_object = get_resource(
+        module, object_list, _get_object_identifier(module),
+        identity_paths=[['id'], ['properties', 'zone_name']],
+    )
 
     if existing_object is None:
         module.exit_json(changed=False)
         return
 
-    existing_object_id_by_new_name = get_resource_id(module, object_list, object_name)
+    existing_object_id_by_new_name = get_resource_id(
+        module, object_list, object_name, identity_paths=[['id'], ['properties', 'zone_name']],
+    )
 
     if (
         existing_object.id is not None
@@ -405,7 +374,10 @@ def update_object(module, client):
 
 
 def remove_object(module, client):
-    existing_object = get_resource(module, _get_object_list(module, client), _get_object_identifier(module))
+    existing_object = get_resource(
+        module, _get_object_list(module, client),
+        _get_object_identifier(module), identity_paths=[['id'], ['properties', 'zone_name']],
+    )
 
     if existing_object is None:
         module.exit_json(changed=False)
@@ -491,12 +463,12 @@ def main():
     module = AnsibleModule(argument_spec=get_module_arguments(), supports_check_mode=True)
 
     if not HAS_SDK:
-        module.fail_json(msg='ionoscloud_container_registry is required for this module, '
-                             'run `pip install ionoscloud_container_registry`')
+        module.fail_json(msg='ionoscloud_dns is required for this module, '
+                             'run `pip install ionoscloud_dns`')
 
 
-    client = ionoscloud_container_registry.ApiClient(get_sdk_config(module, ionoscloud_container_registry))
-    client.user_agent = CONTAINER_REGISTRY_USER_AGENT
+    client = ionoscloud_dns.ApiClient(get_sdk_config(module, ionoscloud_dns))
+    client.user_agent = DNS_USER_AGENT
 
     state = module.params.get('state')